backend/apsi_backend/requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
gunicorn==19.9.0
djangorestframework==3.11.0
Markdown==3.2.1
pyyaml==5.3.1
uritemplate==3.0.1
djangorestframework_simplejwt==4.4.0